NeqE no longer necessary
With the new position of the variables, the energy is in the code in
position 5. Thus neqE (in python there is a shift of one as arrays star
from 0) is always 4, so there is no need for the optional parameter.
However the neutral density is changed from position 5 to 8 if MHD is
enabled, to use with MHD runs use in python th optional parameter
‘neqdyn=8’ (the number of dynamical variables, or equations)